refactor product service

+++ src/main/java/com/mindhub/ms_product/services/Impl/ProductServiceImpl.java
@@ -1,18 +1,27 @@
 package com.mindhub.ms_product.services.Impl;
 
+import com.mindhub.ms_product.config.JwtUtils;
+import com.mindhub.ms_product.exceptions.NotAuthorizedException;
+import com.mindhub.ms_product.exceptions.NotValidArgumentException;
 import com.mindhub.ms_product.repositories.ProductRepository;
 import com.mindhub.ms_product.dtos.ProductDTO;
 import com.mindhub.ms_product.exceptions.NotFoundException;
 import com.mindhub.ms_product.mappers.ProductMapper;
 import com.mindhub.ms_product.models.Product;
 import com.mindhub.ms_product.services.ProductService;
+import jakarta.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.interceptor.TransactionAspectSupport;
 
 import java.util.List;
 
 @Service
 public class ProductServiceImpl implements ProductService {
+
+    private static final Logger log = LoggerFactory.getLogger(ProductServiceImpl.class);
 
     @Autowired
     private ProductRepository productRepository;
@@ -20,53 +29,89 @@ public class ProductServiceImpl implements ProductService {
     @Autowired
     private ProductMapper productMapper;
 
+    @Autowired
+    private JwtUtils jwtUtils;
+
     @Override
     public ProductDTO getProduct(Long id) throws NotFoundException {
         try {
             return productMapper.productToDTO(findById(id));
         } catch (NotFoundException e) {
+            log.warn(e.getMessage());
             throw new NotFoundException(e.getMessage());
         }
     }
 
     @Override
-    public List<ProductDTO> getAllProducts() {
-        return productMapper.productListToDTO(productRepository.findAll());
+    public List<ProductDTO> getAllProducts() throws NotFoundException {
+        try {
+            List<ProductDTO> productList = productMapper.productListToDTO(productRepository.findAll());
+            validateProductListIsEmpty(productList);
+            return productList;
+        } catch (NotFoundException e) {
+            log.warn(e.getMessage());
+            throw new NotFoundException(e.getMessage());
+        }
     }
 
     @Override
-    public Product createProduct(ProductDTO newProduct) {
-        return save(productMapper.productToEntity(newProduct));
+    @Transactional
+    public ProductDTO createProduct(ProductDTO newProduct) throws NotValidArgumentException, NotAuthorizedException {
+        try {
+            validateIsAdmin();
+            validateAlreadyExistsByName(newProduct.getName());
+            Product savedProduct = save(productMapper.productToEntity(newProduct));
+            return productMapper.productToDTO(savedProduct);
+        } catch (NotValidArgumentException | NotAuthorizedException e) {
+            log.warn(e.getMessage());
+            TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+            throw e;
+        }
     }
 
     @Override
-    public Product updateProduct(Long id, ProductDTO updatedProduct) throws NotFoundException {
+    @Transactional
+    public ProductDTO updateProduct(Long id, ProductDTO updatedProduct) throws NotFoundException, NotValidArgumentException, NotAuthorizedException {
+        try {
+            validateIsAdmin();
+            Product productToUpdate = findById(id);
+            validateAlreadyExistsByName(updatedProduct.getName());
+            Product savedProduct = save(productMapper.updateProductToEntity(productToUpdate, updatedProduct));
+            return productMapper.productToDTO(savedProduct);
+        } catch (NotFoundException | NotValidArgumentException | NotAuthorizedException e) {
+            log.warn(e.getMessage());
+            TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+            throw e;
+        }
+    }
+
+    @Override
+    @Transactional
+    public ProductDTO patchProduct(Long id, ProductDTO updatedProduct) throws NotFoundException, NotValidArgumentException {
         try {
             Product productToUpdate = findById(id);
-            return save(productMapper.updateProductToEntity(productToUpdate, updatedProduct));
-        } catch (NotFoundException e) {
-            throw new NotFoundException(e.getMessage());
-        }
-    }
-
-    @Override
-    public Product patchProduct(Long id, ProductDTO updatedProduct) throws NotFoundException {
-        try {
-            Product productToUpdate = findById(id);
-            return save(productMapper.patchProductToEntity(productToUpdate, updatedProduct));
-        } catch (NotFoundException e) {
-            throw new NotFoundException(e.getMessage());
+            validateAlreadyExistsByName(updatedProduct.getName());
+            Product savedProduct = (productMapper.patchProductToEntity(productToUpdate, updatedProduct));
+            return productMapper.productToDTO(savedProduct);
+        } catch (NotFoundException | NotValidArgumentException e) {
+            log.warn(e.getMessage());
+            TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+            throw e;
         }
     }
 
 
     @Override
-    public void deleteProduct(Long id) throws NotFoundException {
+    @Transactional
+    public void deleteProduct(Long id) throws NotFoundException, NotAuthorizedException {
         try {
+            validateIsAdmin();
             existsById(id);
             deleteById(id);
-        } catch (NotFoundException e){
-            throw new NotFoundException(e.getMessage());
+        } catch (NotFoundException | NotAuthorizedException e){
+            log.warn(e.getMessage());
+            TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+            throw e;
         }
     }
 
@@ -94,6 +139,27 @@ public class ProductServiceImpl implements ProductService {
     public void existsById(Long id) throws NotFoundException {
         if (!productRepository.existsById(id)) {
             throw new NotFoundException("Product not found.");
+        }
+    }
+
+    private void validateAlreadyExistsByName(String name) throws NotValidArgumentException {
+        if (productRepository.existsByName(name)) {
+            throw new NotValidArgumentException("Product with name: " + name + "  already exists.");
+        }
+    }
+
+    private void validateProductListIsEmpty(List<ProductDTO> allProducts) throws NotFoundException {
+        if (allProducts.toArray().length == 0) {
+            throw new NotFoundException("No products found to show.");
+        }
+    }
+
+    private void validateIsAdmin() throws NotAuthorizedException {
+        String token = jwtUtils.getJwtToken();
+        String userRole = jwtUtils.extractRole(token);
+
+        if (!userRole.equals("ADMIN")) {
+            throw new NotAuthorizedException("You are not authorized to access this resource.");
         }
     }
 }
